+++ utils/__init__.py
@@ -1,2 +1,12 @@
 from .progress_bar import ProgressBar
-from .read_config import read_config
+from .config import read_config
+from .package import has_package, package_of, package_python_version
+
+
+__all__ = [
+    "ProgressBar",
+    "read_config",
+    "has_package",
+    "package_of",
+    "package_python_version"
+]
